tashas-collection/server/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Cors==3.0.10
psycopg2-binary==2.9.10
python-dotenv==1.0.0
